requirements.txt
# Runtime
oracledb>=2.4.1
httpx[http2]>=0.27.0
apscheduler>=3.10.4
PySide6>=6.10.0
python-dotenv>=1.0.1
cryptography>=42.0.8
pywin32>=306; sys_platform == "win32"

# Servidor web (webhook-only)
fastapi>=0.111.0
uvicorn[standard]>=0.30.1
jinja2>=3.1.4

# Utilitarios
loguru>=0.7.2
tenacity>=8.3.0
pydantic>=2.7.1
pydantic-settings>=2.3.1
